Renaming
* mv instagram_postgres-inserter insta_postgres-inserter
* mv instagram_scraper insta_scraper
* renaming in these files
* renaming in .circleci/config.yml
* renaming in docker-compose.yml
