+++ .circleci/config.yml
@@ -14,83 +14,83 @@ jobs:
       - checkout
       # specify any bash command here prefixed with `run: `
       - run: go test -v ./...
-  scraper-image:
+  insta_scraper-image:
     # specify the version
     executor: job-runner
     environment:
-      SCRAPER_IMAGE_NAME: codesmag/scraper
+      INSTA_SCRAPER_IMAGE_NAME: codesmag/insta_scraper
     steps:
       - checkout
       - setup_remote_docker
       - run:
           name: Build Docker Image
-          command: docker build -f scraper/Dockerfile -t $SCRAPER_IMAGE_NAME .
+          command: docker build -f insta_scraper/Dockerfile -t $INSTA_SCRAPER_IMAGE_NAME .
       - run:
           name: Publish Docker Image to Docker Hub
           command: |
             echo "$DOCKERHUB_PASS" | docker login -u "$DOCKERHUB_USERNAME" --password-stdin
             IMAGE_TAG=$(date "+%Y-%m-%dT%H-%M-%S")
-            docker tag $SCRAPER_IMAGE_NAME:latest $SCRAPER_IMAGE_NAME:$IMAGE_TAG
-            docker push $SCRAPER_IMAGE_NAME:latest
-            docker push $SCRAPER_IMAGE_NAME:$IMAGE_TAG
-  postgres-inserter-image:
+            docker tag $INSTA_SCRAPER_IMAGE_NAME:latest $INSTA_SCRAPER_IMAGE_NAME:$IMAGE_TAG
+            docker push $INSTA_SCRAPER_IMAGE_NAME:latest
+            docker push $INSTA_SCRAPER_IMAGE_NAME:$IMAGE_TAG
+  insta_postgres-inserter-image:
     # specify the version
     executor: job-runner
     environment:
-      SCRAPER_IMAGE_NAME: codesmag/postgres-inserter
+      INSTA_POSTGRES_INSERTER_IMAGE_NAME: codesmag/insta_postgres-inserter
     steps:
       - checkout
       - setup_remote_docker
       - run:
           name: Build Docker Image
-          command: docker build -f postgres-inserter/Dockerfile -t $SCRAPER_IMAGE_NAME .
+          command: docker build -f insta_postgres-inserter/Dockerfile -t $POSTGRES_INSERTER_IMAGE_NAME .
       - run:
           name: Publish Docker Image to Docker Hub
           command: |
             echo "$DOCKERHUB_PASS" | docker login -u "$DOCKERHUB_USERNAME" --password-stdin
             IMAGE_TAG=$(date "+%Y-%m-%dT%H-%M-%S")
-            docker tag $SCRAPER_IMAGE_NAME:latest $SCRAPER_IMAGE_NAME:$IMAGE_TAG
-            docker push $SCRAPER_IMAGE_NAME:latest
-            docker push $SCRAPER_IMAGE_NAME:$IMAGE_TAG
+            docker tag $POSTGRES_INSERTER_IMAGE_NAME:latest $POSTGRES_INSERTER_IMAGE_NAME:$IMAGE_TAG
+            docker push $POSTGRES_INSERTER_IMAGE_NAME:latest
+            docker push $POSTGRES_INSERTER_IMAGE_NAME:$IMAGE_TAG
 
   neo4j-inserter-image:
     # specify the version
     executor: job-runner
     environment:
-      SCRAPER_IMAGE_NAME: codesmag/neo4j-inserter
+      NEO4J_INSERTER_IMAGE_NAME: codesmag/neo4j-inserter
     steps:
       - checkout
       - setup_remote_docker
       - run:
           name: Build Docker Image
-          command: docker build -f neo4j-inserter/Dockerfile -t $SCRAPER_IMAGE_NAME .
+          command: docker build -f neo4j-inserter/Dockerfile -t $NEO4J_INSERTER_IMAGE_NAME .
       - run:
           name: Publish Docker Image to Docker Hub
           command: |
             echo "$DOCKERHUB_PASS" | docker login -u "$DOCKERHUB_USERNAME" --password-stdin
             IMAGE_TAG=$(date "+%Y-%m-%dT%H-%M-%S")
-            docker tag $SCRAPER_IMAGE_NAME:latest $SCRAPER_IMAGE_NAME:$IMAGE_TAG
-            docker push $SCRAPER_IMAGE_NAME:latest
-            docker push $SCRAPER_IMAGE_NAME:$IMAGE_TAG
+            docker tag $NEO4J_INSERTER_IMAGE_NAME:latest $NEO4J_INSERTER_IMAGE_NAME:$IMAGE_TAG
+            docker push $NEO4J_INSERTER_IMAGE_NAME:latest
+            docker push $NEO4J_INSERTER_IMAGE_NAME:$IMAGE_TAG
   dgraph-inserter-image:
     # specify the version
     executor: job-runner
     environment:
-      SCRAPER_IMAGE_NAME: codesmag/dgraph-inserter
+      DGRAPH_INSERTER_IMAGE_NAME: codesmag/dgraph-inserter
     steps:
       - checkout
       - setup_remote_docker
       - run:
           name: Build Docker Image
-          command: docker build -f dgraph-inserter/Dockerfile -t $SCRAPER_IMAGE_NAME .
+          command: docker build -f dgraph-inserter/Dockerfile -t $DGRAPH_INSERTER_IMAGE_NAME .
       - run:
           name: Publish Docker Image to Docker Hub
           command: |
             echo "$DOCKERHUB_PASS" | docker login -u "$DOCKERHUB_USERNAME" --password-stdin
             IMAGE_TAG=$(date "+%Y-%m-%dT%H-%M-%S")
-            docker tag $SCRAPER_IMAGE_NAME:latest $SCRAPER_IMAGE_NAME:$IMAGE_TAG
-            docker push $SCRAPER_IMAGE_NAME:latest
-            docker push $SCRAPER_IMAGE_NAME:$IMAGE_TAG
+            docker tag $DGRAPH_INSERTER_IMAGE_NAME:latest $DGRAPH_INSERTER_IMAGE_NAME:$IMAGE_TAG
+            docker push $DGRAPH_INSERTER_IMAGE_NAME:latest
+            docker push $DGRAPH_INSERTER_IMAGE_NAME:$IMAGE_TAG
 
 workflows:
   version: 2
