+++ insta_scraper/main/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	scraper "github.com/codeuniversity/smag-mvp/instagram_scraper"
+	scraper "github.com/codeuniversity/smag-mvp/insta_scraper"
 	"github.com/codeuniversity/smag-mvp/kafka"
 	"github.com/codeuniversity/smag-mvp/service"
 )
